ABI changed, code formatted/normalized.
all cases of `extern "C"` changed to `extern "C-unwind"`.
This may or may not be unsound, but tee hee! it works, and
now panics can fully unwind and give useful info.

libunwind gets angry at not finding an exception handler with the right
personality before the bottom of the stack, but tee hee! it just aborts
after producing a backtrace. works for me.

Signed-off-by: Julia DeMille <me@jdemille.com>